+++ Cron/run_task.sh
@@ -4,7 +4,7 @@ VIRTUAL_ENV="/opt/gvm/gvm"
 SCRIPT_PATH="/opt/gvm/Targets_Tasks/run-task.py"
 
 source "$VIRTUAL_ENV/bin/activate"
-python3 "$SCRIPT_PATH"
+python3 "$SCRIPT_PATH" "$@"
 
 deactivate